Broke LocalItemsFile.xml into multiple files
The regular LocalItemsFile was too large to be read in by Windows Phone
and would cause memory crashes. As a result, this file was broken down
into six files, one for each generation. LocalItemFileService was
updated to support reading in multiple files. Parser was also updated to
break up XML files by generation.